+++ test/unit/elastic_class/ElasticSearchRepo_delete_dump.py
@@ -36,57 +36,6 @@ import version
 __version__ = version.__version__
 
 
-class Repo(object):
-
-    """Class:  Repo
-
-    Description:  Class representation of the snapshot class.
-
-    Methods:
-        get_repository -> Stub holder for snapshot.get_repository method.
-        create_repository -> Stub holder for snapshot.create_repository method.
-        delete_repository -> Stub holder for snapshot.delete_repository method.
-
-    """
-
-    def delete_repository(self):
-
-        """Method:  delete_repository
-
-        Description:  Stub holder for snapshot.delete_repository method.
-
-        Arguments:
-
-        """
-
-        return {"acknowledged": True}
-
-    def create_repository(self):
-
-        """Method:  create_repository
-
-        Description:  Stub holder for snapshot.create_repository method.
-
-        Arguments:
-
-        """
-
-        return {"acknowledged": True}
-
-    def get_repository(self):
-
-        """Method:  get_repository
-
-        Description:  Stub holder for snapshot.get_repository method.
-
-        Arguments:
-
-        """
-
-        return {"reponame": {"type": "dbdump", "settings":
-                             {"location": "/dir/path/dump"}}}
-
-
 class Elasticsearch(object):
 
     """Class:  ElasticSearch
@@ -95,8 +44,6 @@ class Elasticsearch(object):
 
     Methods:
         __init__ -> Initialize configuration environment.
-        ping -> Stub holder for Elasticsearch.ping method.
-        info -> Stub holder for Elasticsearch.info method.
 
     """
 
@@ -115,31 +62,6 @@ class Elasticsearch(object):
         self.ping_status = True
         self.info_status = {"cluster_name": "ClusterName",
                             "name": "servername"}
-        self.snapshot = Repo()
-
-    def ping(self):
-
-        """Method:  ping
-
-        Description:  Stub holder for Elasticsearch.ping method.
-
-        Arguments:
-
-        """
-
-        return self.ping_status
-
-    def info(self):
-
-        """Method:  info
-
-        Description:  Stub holder for Elasticsearch.info method.
-
-        Arguments:
-
-        """
-
-        return self.info_status
 
 
 class UnitTest(unittest.TestCase):
@@ -174,25 +96,18 @@ class UnitTest(unittest.TestCase):
         self.repo2 = "reponame2"
         self.es = Elasticsearch(self.host_list)
         self.dump_name = "dump3"
-        self.nodes_data = {"serverid1": {"name": "hostname1", "settings":
-            {"path": {"data": ["/dir/data1"], "logs": ["/dir/logs1"]}}},
-            "serverid2": {"name": "hostname2", "settings":
-            {"path": {"data": ["/dir/data2"], "logs": ["/dir/logs2"]}}}}
-        self.health_data = {"status": "green", "cluster_name": "ClusterName"}
+        self.repo_dict = {"reponame": {"type": "dbdump", "settings":
+                                       {"location": "/dir/path/dump"}}}
 
     @mock.patch("elastic_class.delete_snapshot",
                 mock.Mock(return_value={"acknowledged": False}))
-    @mock.patch("elastic_class.get_cluster_nodes",
-                mock.Mock(return_value={"_nodes": {"total": 3}}))
-    @mock.patch("elastic_class.get_master_name",
-                mock.Mock(return_value="MasterName"))
-    @mock.patch("elastic_class.get_info",
-                mock.Mock(return_value={"name": "localservername"}))
-    @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_nodes")
+    @mock.patch("elastic_class.ElasticSearch.update_status",
+                mock.Mock(return_value=True))
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_repo_name_none(self, mock_es, mock_list, mock_nodes, mock_health):
+    def test_repo_name_none(self, mock_es, mock_list, mock_repo):
 
         """Function:  test_repo_name_none
 
@@ -205,28 +120,24 @@ class UnitTest(unittest.TestCase):
         mock_es.return_value = self.es
         mock_list.side_effect = [[["dump1"], ["dump2"], ["dump3"]],
                                  [["dump1"], ["dump2"]]]
-        mock_nodes.return_value = self.nodes_data
-        mock_health.return_value = self.health_data
+        mock_repo.return_value = self.repo_dict
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo)
         es.repo = None
-        self.assertEqual(es.delete_dump(dump_name=self.dump_name),
+        self.assertEqual(
+            es.delete_dump(dump_name=self.dump_name),
             (True,
              "ERROR: Missing arg/repo not exist, Repo: None, Dump: dump3"))
 
     @mock.patch("elastic_class.delete_snapshot",
                 mock.Mock(return_value={"acknowledged": False}))
-    @mock.patch("elastic_class.get_cluster_nodes",
-                mock.Mock(return_value={"_nodes": {"total": 3}}))
-    @mock.patch("elastic_class.get_master_name",
-                mock.Mock(return_value="MasterName"))
-    @mock.patch("elastic_class.get_info",
-                mock.Mock(return_value={"name": "localservername"}))
-    @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_nodes")
+    @mock.patch("elastic_class.ElasticSearch.update_status",
+                mock.Mock(return_value=True))
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_no_dump(self, mock_es, mock_list, mock_nodes, mock_health):
+    def test_no_dump(self, mock_es, mock_list, mock_repo):
 
         """Function:  test_no_dump
 
@@ -239,26 +150,22 @@ class UnitTest(unittest.TestCase):
         mock_es.return_value = self.es
         mock_list.side_effect = [[["dump1"], ["dump2"]],
                                  [["dump1"], ["dump2"]]]
-        mock_nodes.return_value = self.nodes_data
-        mock_health.return_value = self.health_data
+        mock_repo.return_value = self.repo_dict
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo)
-        self.assertEqual(es.delete_dump(self.repo, self.dump_name),
+        self.assertEqual(
+            es.delete_dump(self.repo, self.dump_name),
             (True, "ERROR: Dump: dump3 not in Repository: reponame"))
 
     @mock.patch("elastic_class.delete_snapshot",
                 mock.Mock(return_value={"acknowledged": True}))
-    @mock.patch("elastic_class.get_cluster_nodes",
-                mock.Mock(return_value={"_nodes": {"total": 3}}))
-    @mock.patch("elastic_class.get_master_name",
-                mock.Mock(return_value="MasterName"))
-    @mock.patch("elastic_class.get_info",
-                mock.Mock(return_value={"name": "localservername"}))
-    @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_nodes")
+    @mock.patch("elastic_class.ElasticSearch.update_status",
+                mock.Mock(return_value=True))
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_dump_detected(self, mock_es, mock_list, mock_nodes, mock_health):
+    def test_dump_detected(self, mock_es, mock_list, mock_repo):
 
         """Function:  test_dump_detected
 
@@ -271,26 +178,22 @@ class UnitTest(unittest.TestCase):
         mock_es.return_value = self.es
         mock_list.side_effect = [[["dump1"], ["dump2"], ["dump3"]],
                                  [["dump1"], ["dump2"], ["dump3"]]]
-        mock_nodes.return_value = self.nodes_data
-        mock_health.return_value = self.health_data
+        mock_repo.return_value = self.repo_dict
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo)
-        self.assertEqual(es.delete_dump(self.repo, self.dump_name),
+        self.assertEqual(
+            es.delete_dump(self.repo, self.dump_name),
             (True, "ERROR: Dump still detected: reponame, dump3"))
 
     @mock.patch("elastic_class.delete_snapshot",
                 mock.Mock(return_value={"acknowledged": False}))
-    @mock.patch("elastic_class.get_cluster_nodes",
-                mock.Mock(return_value={"_nodes": {"total": 3}}))
-    @mock.patch("elastic_class.get_master_name",
-                mock.Mock(return_value="MasterName"))
-    @mock.patch("elastic_class.get_info",
-                mock.Mock(return_value={"name": "localservername"}))
-    @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_nodes")
+    @mock.patch("elastic_class.ElasticSearch.update_status",
+                mock.Mock(return_value=True))
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_delete_failed(self, mock_es, mock_list, mock_nodes, mock_health):
+    def test_delete_failed(self, mock_es, mock_list, mock_repo):
 
         """Function:  test_delete_failed
 
@@ -303,27 +206,23 @@ class UnitTest(unittest.TestCase):
         mock_es.return_value = self.es
         mock_list.side_effect = [[["dump1"], ["dump2"], ["dump3"]],
                                  [["dump1"], ["dump2"]]]
-        mock_nodes.return_value = self.nodes_data
-        mock_health.return_value = self.health_data
+        mock_repo.return_value = self.repo_dict
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo)
         es.repo_dict[self.repo2] = True
-        self.assertEqual(es.delete_dump(self.repo2, self.dump_name),
+        self.assertEqual(
+            es.delete_dump(self.repo2, self.dump_name),
             (True, "ERROR:  Dump deletion failed:  reponame2, dump3"))
 
     @mock.patch("elastic_class.delete_snapshot",
                 mock.Mock(return_value={"acknowledged": True}))
-    @mock.patch("elastic_class.get_cluster_nodes",
-                mock.Mock(return_value={"_nodes": {"total": 3}}))
-    @mock.patch("elastic_class.get_master_name",
-                mock.Mock(return_value="MasterName"))
-    @mock.patch("elastic_class.get_info",
-                mock.Mock(return_value={"name": "localservername"}))
-    @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_nodes")
+    @mock.patch("elastic_class.ElasticSearch.update_status",
+                mock.Mock(return_value=True))
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_no_repo_name(self, mock_es, mock_list, mock_nodes, mock_health):
+    def test_no_repo_name(self, mock_es, mock_list, mock_repo):
 
         """Function:  test_no_repo_name
 
@@ -336,8 +235,7 @@ class UnitTest(unittest.TestCase):
         mock_es.return_value = self.es
         mock_list.side_effect = [[["dump1"], ["dump2"], ["dump3"]],
                                  [["dump1"], ["dump2"]]]
-        mock_nodes.return_value = self.nodes_data
-        mock_health.return_value = self.health_data
+        mock_repo.return_value = self.repo_dict
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo)
         self.assertEqual(es.delete_dump(dump_name=self.dump_name),
@@ -345,17 +243,13 @@ class UnitTest(unittest.TestCase):
 
     @mock.patch("elastic_class.delete_snapshot",
                 mock.Mock(return_value={"acknowledged": True}))
-    @mock.patch("elastic_class.get_cluster_nodes",
-                mock.Mock(return_value={"_nodes": {"total": 3}}))
-    @mock.patch("elastic_class.get_master_name",
-                mock.Mock(return_value="MasterName"))
-    @mock.patch("elastic_class.get_info",
-                mock.Mock(return_value={"name": "localservername"}))
-    @mock.patch("elastic_class.get_cluster_health")
-    @mock.patch("elastic_class.get_nodes")
+    @mock.patch("elastic_class.ElasticSearch.update_status",
+                mock.Mock(return_value=True))
+    @mock.patch("elastic_class.is_active", mock.Mock(return_value=True))
+    @mock.patch("elastic_class.get_repo_list")
     @mock.patch("elastic_class.get_dump_list")
     @mock.patch("elastic_class.elasticsearch.Elasticsearch")
-    def test_default(self, mock_es, mock_list, mock_nodes, mock_health):
+    def test_default(self, mock_es, mock_list, mock_repo):
 
         """Function:  test_default
 
@@ -368,8 +262,7 @@ class UnitTest(unittest.TestCase):
         mock_es.return_value = self.es
         mock_list.side_effect = [[["dump1"], ["dump2"], ["dump3"]],
                                  [["dump1"], ["dump2"]]]
-        mock_nodes.return_value = self.nodes_data
-        mock_health.return_value = self.health_data
+        mock_repo.return_value = self.repo_dict
 
         es = elastic_class.ElasticSearchRepo(self.host_list, repo=self.repo)
         self.assertEqual(es.delete_dump(self.repo, self.dump_name),